backend/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
ecdsa==0.19.0
fastapi==0.110.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.2.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
